refactor: extract Jinja2 filters into filters.py and clean up app factory
